file cleanup; PR feedback
