+++ csharp/src/Apache.Arrow.Adbc.FlightSql/Apache.Arrow.Adbc.FlightSql.csproj
@@ -6,7 +6,7 @@
 
   <ItemGroup>
     <PackageReference Include="Apache.Arrow.Flight" Version="12.0.0" />
-    <PackageReference Include="System.Net.Http.WinHttpHandler" Version="7.0.0" />
+    <PackageReference Include="System.Net.Http.WinHttpHandler" Version="7.0.0" Condition="'$(TargetFrameworkIdentifier)' == '.NETStandard'" />
   </ItemGroup>
 
   <ItemGroup>
