+++ csharp/src/Apache.Arrow.Adbc.FlightSql/FlightSqlStatement.cs
@@ -16,8 +16,8 @@
  */
 
 using System;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using Apache.Arrow.Adbc.Core;
 using Apache.Arrow.Flight;
 using Grpc.Core;
 
@@ -186,6 +186,34 @@ namespace Apache.Arrow.Adbc.FlightSql
             // -- union array
 
             return null;
+        }
+
+        /// <summary>
+        /// For decimals, Arrow throws an OverflowException if a value is < decimal.min or > decimal.max
+        /// So parse the numeric value and return it as a string, if possible
+        /// </summary>
+        /// <param name="oex"></param>
+        /// <returns>A string value of the decimal that threw the exception or rethrows the OverflowException.</returns>
+        /// <exception cref="ArgumentNullException"></exception>
+        private string ParseDecimalValueFromOverflowException(OverflowException oex)
+        {
+            if (oex == null)
+                throw new ArgumentNullException(nameof(oex));
+
+            // any decimal value, positive or negative, with or without a decimal in place
+            Regex regex = new Regex(" -?\\d*\\.?\\d* ");
+
+            var matches = regex.Matches(oex.Message);
+
+            foreach (Match match in matches)
+            {
+                string value = match.Value;
+
+                if (!string.IsNullOrEmpty(value))
+                    return value;
+            }
+
+            throw oex;
         }
     }
 }
