+++ csharp/src/Apache.Arrow.Adbc/Interop/NativeDelegate.cs
